Add new required freq argument.
Supplied freq should be the center frequency of the AP. The
freq will then be used to set the scan_freq and freq_list
options of the supplicant conf. The idea being to optimize
the periodic scanning of the supplicant to reduce the amount
of down time experienced. This will also help in cases where
another AP exists that matches the same ssid and bssid on
a different frequency.